// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  admin: 'admin',
  employee: 'employee'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const VehicleType = {
  TwoWheeler: 'TwoWheeler',
  FourWheeler: 'FourWheeler',
  Commercial: 'Commercial',
  Other: 'Other'
} as const

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType]